django_shop/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
Pillow==9.0.0

# development
flake8
python-lsp-server
autopep8
loguru
